refactor(post): move decidePublishedAt to shared/ (dep direction)
shared/post/list.ts imported decidePublishedAt from @post/shared/, i.e. a
shared/ module depending on a domains/ module — a violation of the
unidirectional dependency rule (routes -> domains -> shared -> core).

Relocate the helper and its co-located test from domains/post/shared/ to
shared/post/ (it now has consumers in both shared/ and domains/, so shared/
is its correct home) and repoint the two importers (shared/post/list.ts and
PostDetails.handler.ts) to @shared/post/decide_published_at.ts. Every edge is
now shared->shared or domains->shared. Pure relocation; no behavior change.
